Ported to Debian 9 with libgtk 2.24.31. Removed some warnings about type convert from 32bit to 64bit.

+++ iup/src/gtk/iupgtk_open.c
@@ -237,7 +237,7 @@ static void gtkSetGlobalAttrib(void)
   GdkDisplay* display = gdk_display_get_default();
   Display* xdisplay = GDK_DISPLAY_XDISPLAY(display);
   IupSetGlobal("XDISPLAY", (char*)xdisplay);
-  IupSetGlobal("XSCREEN", (char*)XDefaultScreen(xdisplay));
+  IupSetGlobal("XSCREEN", (char*)((long)XDefaultScreen(xdisplay)));
   IupSetGlobal("XSERVERVENDOR", ServerVendor(xdisplay));
   IupSetInt(NULL, "XVENDORRELEASE", VendorRelease(xdisplay));
 }
